add e2e test skeleton
