+++ e2e/test/scenarios/embedding-sdk/helpers/save-interactive-question-e2e-helpers.ts
@@ -1,44 +1,4 @@
-import {
-  entityPickerModal,
-  modal,
-  popover,
-  visitFullAppEmbeddingUrl,
-} from "e2e/support/helpers";
-import { EMBEDDING_SDK_STORY_HOST } from "e2e/support/helpers/e2e-embedding-sdk-helpers";
-import { JWT_SHARED_SECRET } from "e2e/support/helpers/e2e-jwt-helpers";
-
-/** Get storybook args in the format of as "key:value;key:value" */
-export function getStorybookArgs(props: Record<string, string>): string {
-  const params = new URLSearchParams(props);
-
-  return params.toString().replaceAll("=", ":").replaceAll("&", ";");
-}
-
-export function visitInteractiveQuestionStory(
-  options: { saveToCollectionId?: number } = {},
-) {
-  const params: Record<string, string> = {
-    ...(options.saveToCollectionId && {
-      saveToCollectionId: options.saveToCollectionId.toString(),
-    }),
-  };
-
-  cy.get("@questionId").then(questionId => {
-    visitFullAppEmbeddingUrl({
-      url: EMBEDDING_SDK_STORY_HOST,
-      qs: {
-        id: "embeddingsdk-interactivequestion--default",
-        viewMode: "story",
-        args: getStorybookArgs(params),
-      },
-      onBeforeLoad: (window: any) => {
-        window.JWT_SHARED_SECRET = JWT_SHARED_SECRET;
-        window.METABASE_INSTANCE_URL = Cypress.config().baseUrl;
-        window.QUESTION_ID = questionId;
-      },
-    });
-  });
-}
+import { entityPickerModal, modal, popover } from "e2e/support/helpers";
 
 export function saveInteractiveQuestionAsNewQuestion(options: {
   questionName: string;
